lowered frame rate various optimizations - less divisions etc, slightly faster updates

+++ projects/cartuchnik/loader/src/pecado.h
@@ -9,7 +9,7 @@ extern "C" {
 int8_t isin(uint8_t angle);
 int8_t icos(uint8_t angle);
 void irotate(int16_t* x, int16_t* y, int16_t ox, int16_t oy, uint8_t angle);
-void irotate0(int16_t* x, int16_t* y, uint8_t angle);
+void irotate0(int8_t* x, int8_t* y, uint8_t angle);
 
 #ifdef __cplusplus
 }
